Fixed the connection problem with the server on the map page

+++ lib/features/visitor/data/models/request/set_location_request.dart
@@ -1,8 +1,8 @@
 class SetLocationRequest {
   int? userId;
   int? checkPointId;
-  int? lat;
-  int? lng;
+  double? lat;
+  double? lng;
   bool? report;
   List<Attachments>? attachments;
